chore: Move HyperKZG+Shponk code to hyperkzg.rs source file
