+++ benches/pcs.rs
@@ -1,10 +1,8 @@
-use arecibo::provider::Bn256EngineIPA;
 use arecibo::provider::{
   hyperkzg::EvaluationEngine as MLEvaluationEngine,
-  hyperkzg_shplonk::EvaluationEngine as MLShplEvaluationEngine,
   ipa_pc::EvaluationEngine as IPAEvaluationEngine, non_hiding_zeromorph::ZMPCS,
-  shplonk::EvaluationEngine as Shplonk, Bn256EngineKZG, Bn256EngineZM,
 };
+use arecibo::provider::{Bn256EngineIPA, Bn256EngineKZG, Bn256EngineZM};
 use arecibo::spartan::polys::multilinear::MultilinearPolynomial;
 use arecibo::traits::{
   commitment::CommitmentEngineTrait, evaluation::EvaluationEngineTrait, Engine,
@@ -161,9 +159,7 @@ fn bench_pcs(c: &mut Criterion) {
     bench_pcs_verifying_internal,
     (ipa_assets, IPAEvaluationEngine<Bn256EngineIPA>),
     (hyperkzg_assets, MLEvaluationEngine<Bn256, Bn256EngineKZG>),
-    (zm_assets, ZMPCS<Bn256, Bn256EngineZM>),
-    (shplonk_assets, Shplonk<Bn256, Bn256EngineKZG>),
-    (hyperkzg_shplonk_assets, MLShplEvaluationEngine<Bn256, Bn256EngineKZG>)
+    (zm_assets, ZMPCS<Bn256, Bn256EngineZM>)
   );
 }
