+++ src/provider/mod.rs
@@ -4,8 +4,6 @@
 pub mod hyperkzg;
 pub mod ipa_pc;
 pub mod non_hiding_zeromorph;
-pub mod shplonk;
-pub mod hyperkzg_shplonk;
 
 // crate-public modules, made crate-public mostly for tests
 pub(crate) mod bn256_grumpkin;
